Rag Loan Chatbot/requirement.txt
```txt
pandas
faiss-cpu
sentence-transformers
openai
streamlit
transformers
torch
```